Remove extra files and simplify code
