+++ package.json
@@ -85,7 +85,6 @@
     "react-i18next": "^13.2.1",
     "react-load-script": "^0.0.6",
     "react-markdown": "^8.0.7",
-    "react-player": "^2.12.0",
     "react-query": "^3.39.3",
     "react-slider": "^2.0.5",
     "react-syntax-highlighter": "^15.5.0",
@@ -115,6 +114,6 @@
     "@types/styled-components": "^5.1.26",
     "@types/uuid": "^9.0.1",
     "prettier": "^2.8.8",
-    "prisma": "^5.2.0"
+    "prisma": "^4.13.0"
   }
 }
