+++ src/components/shorts/ImageType.ts
@@ -1,0 +1,8 @@
+export interface ImageType {
+  id: string;
+  image: string;
+  title: string;
+  topic: string;
+  link: string;
+  onClick?: () => void;
+}
