+++ src/components/shorts/sampleVideos.ts
@@ -1,0 +1,39 @@
+import { ImageType } from '@/components/shorts/ImageType';
+
+export const videos: ImageType[] = [
+  {
+    id: '1',
+    image: 'https://i.ytimg.com/vi/NHxjypDc4xw/oar1.jpg',
+    title: 'What is leverage',
+    topic: 'Blockchain',
+    link: 'https://www.youtube.com/shorts/NHxjypDc4xw',
+  },
+  {
+    id: '2',
+    image: 'https://i.ytimg.com/vi/vEyNb7D5gb0/oar1.jpg',
+    title: 'What is leverage',
+    topic: 'Blockchain',
+    link: 'https://www.youtube.com/shorts/vEyNb7D5gb0',
+  },
+  {
+    id: '3',
+    image: 'https://i.ytimg.com/vi/vEyNb7D5gb0/oar1.jpg',
+    title: 'What is leverage',
+    topic: 'Blockchain',
+    link: 'https://www.youtube.com/shorts/NHxjypDc4xw',
+  },
+  {
+    id: '4',
+    image: 'https://i.ytimg.com/vi/vEyNb7D5gb0/oar1.jpg',
+    title: 'What is leverage',
+    topic: 'Blockchain',
+    link: 'https://www.youtube.com/shorts/NHxjypDc4xw',
+  },
+  {
+    id: '5',
+    image: 'https://i.ytimg.com/vi/vEyNb7D5gb0/oar1.jpg',
+    title: 'What is leverage',
+    topic: 'Blockchain',
+    link: 'https://www.youtube.com/shorts/NHxjypDc4xw',
+  },
+];
